try_single_request()
function to attempt parsing a request from data currently accumulated in
the incoming buffer for a client connection. this takes place
immediately after reading into the buffer. if a request is not yet
complete in the buffer, program can wait to read more to the buffer in
next event loop iteration before reattempt