non-blocking read/write
functions written to handle the read and write without blocking

the functions perform the nb syscalls and handle unwanted behaviour with
the reads/writes in the loop. they also append to / consume from the
::incoming and ::outgoing buffers for each connection object. they
change the intention flags (want_read, want_write) depending on current
state to keep the loop up to speed on whats going on.