pipelining
try_single_request() now executed in a loop while there are complete
messages present in the buffer. also ensuring that ::incoming is
consumed by the size of the the incoming messsage in current trial loop,
not just dumping the buffer. well, i was never just dumping the buffer,
but its clearer now why that is useful.